Add K8sServices as a sinkable resource
