+++ frontend/packages/knative-plugin/locales/en/knative-plugin.json
@@ -67,6 +67,12 @@
   "The data posted to the target function": "The data posted to the target function",
   "Schedule": "Schedule",
   "Schedule is described using the unix-cron string format (* * * * *)": "Schedule is described using the unix-cron string format (* * * * *)",
+  "This resource will be the sink for the Event source.": "This resource will be the sink for the Event source.",
+  "No resources available": "No resources available",
+  "Select the URI option, or exit this form and create a Knative Service, Broker, or Channel first.": "Select the URI option, or exit this form and create a Knative Service, Broker, or Channel first.",
+  "Select resource": "Select resource",
+  "A Universal Resource Indicator where events are going to be delivered. Ex. \"http://cluster.example.com/svc\"": "A Universal Resource Indicator where events are going to be delivered. Ex. \"http://cluster.example.com/svc\"",
+  "Enter URI": "Enter URI",
   "Enable": "Enable",
   "Password": "Password",
   "CA certificate": "CA certificate",
@@ -94,12 +100,6 @@
   "Match type": "Match type",
   "Subject Name": "Subject Name",
   "Add values": "Add values",
-  "A Universal Resource Indicator where events are going to be delivered. Ex. \"http://cluster.example.com/svc\"": "A Universal Resource Indicator where events are going to be delivered. Ex. \"http://cluster.example.com/svc\"",
-  "Enter URI": "Enter URI",
-  "This resource will be the sink for the Event source.": "This resource will be the sink for the Event source.",
-  "No resources available": "No resources available",
-  "Select the URI option, or exit this form and create a Knative Service, Broker, or Channel first.": "Select the URI option, or exit this form and create a Knative Service, Broker, or Channel first.",
-  "Select resource": "Select resource",
   "Sink": "Sink",
   "Add a sink to route this Event source to a Channel, Broker, Knative Service or another route.": "Add a sink to route this Event source to a Channel, Broker, Knative Service or another route.",
   "Please enter a URI that is less then 2000 characters.": "Please enter a URI that is less then 2000 characters.",
