+++ frontend/packages/knative-plugin/src/components/sink-source/__tests__/SinkSourceModal.spec.tsx
@@ -7,7 +7,7 @@ import {
 } from '@console/internal/components/factory/modal';
 import { formikFormProps } from '@console/shared/src/test-utils/formik-props-utils';
 import { ServiceModel } from '../../../models';
-import { SinkUriResourcesGroup } from '../../add/event-sources/SinkSection';
+import SinkUriResourcesGroup from '../../add/event-sources/form-fields/SinkUriResourcesGroup';
 import SinkSourceModal from '../SinkSourceModal';
 
 type SinkSourceModalProps = React.ComponentProps<typeof SinkSourceModal>;
